Move mock tunnel code to test directory
